Difficulty Update
this update adds 4 new special squares to the game. The squares include the armored square which takes a lot to go down., the vortex which packs a punch, the thief which has sticky finger and the scorer who does not like to play fair. This update also includes bug fixes like the issue where gold squares could not hurt you.