Menu Update
there now is a main menu, game over screen and a help menu. also fixed an issue with the scaling difficulty not working all the time. Also nerfed how effective the remove 1 enemy slot so now it only will remove an enemy if the limit is greater than or equal to 3